Add get comments from photo function

+++ api/country.go
@@ -8,7 +8,7 @@ import (
 
 // Country entity definition
 type Country struct {
-	ID   uint `gorm:"primary_key"`
+	ID   uint64 `gorm:"primary_key"`
 	Name string
 }
